tests/wpt/web-platform-tests/app-history/navigate/navigate-transitionWhile-reject.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async t => {
  let promise_value = { abc: 'def' };
  let promise = Promise.reject(promise_value);
  appHistory.onnavigate = e => e.transitionWhile(promise);
  let returned_promise = appHistory.navigate("#1");
  assert_not_equals(promise, returned_promise);
  return promise_rejects_exactly(t, promise_value, returned_promise);
}, "navigate() and transitionWhile() with a rejected promise");
</script>
